feat: complete refactor
