+++ .eslintignore
@@ -3,3 +3,4 @@
 /.vscode
 /.git
 /test
+/src/js/z_temp
