+++ src/js/app.js
@@ -11,11 +11,7 @@ import setAsideControls from './utility/setAsideControls';
 // display intro & menu
 intro.init();
 
-// display game UI
+// load and display base game UI
 document.body.prepend(displayPlayerTable.init());
-
-// add event listeners to hand controls
 setHandControls();
-
-// add event listeners and animations to aside buttons
 setAsideControls();
