+++ src/js/cards/renderCard.js
@@ -27,7 +27,7 @@ function getRenderCard() {
         if (cardElement.children[0].classList.contains(`card__color--${stack.id}`)) {
           stack.classList.remove('active-zone__stack--empty');
           targetStack.dom = stack;
-          stack.style.width = null;
+          targetStack.dom.style = null;
           targetStack.width = stack.offsetWidth;
           targetStack.height = stack.offsetHeight;
           targetStack.shift = gameState.activePlayer.activeDecks[targetStack.dom.id].shift;
@@ -55,6 +55,7 @@ function getRenderCard() {
           break;
 
         case 'left':
+          console.log('we are in case left');
           while (targetStack.dom.parentElement.offsetWidth / 2.5 < targetStack.width
             + (targetStack.length * cardShiftValue)
             && cardShiftValue !== 10) {
@@ -63,9 +64,9 @@ function getRenderCard() {
           Array.from(targetStack.dom.children).forEach((card, i) => {
             card.style.right = `${i * cardShiftValue}px`;
           });
-          cardElement.style.right = `${targetStack.length * cardShiftValue}px`;
-          if (targetStack.length !== 0) {
-            targetStack.dom.style.width = `${targetStack.width + cardShiftValue * targetStack.length}px`;
+          cardElement.style.right = `${(targetStack.length - 1) * cardShiftValue}px`;
+          if (targetStack.length > 1) {
+            targetStack.dom.style.width = `${targetStack.width + cardShiftValue * (targetStack.length - 1)}px`;
           }
           break;
 
@@ -78,9 +79,9 @@ function getRenderCard() {
           Array.from(targetStack.dom.children).forEach((card, i) => {
             card.style.left = `${i * cardShiftValue}px`;
           });
-          cardElement.style.left = `${targetStack.length * cardShiftValue}px`;
-          if (targetStack.length !== 0) {
-            targetStack.dom.style.width = `${targetStack.width + cardShiftValue * targetStack.length}px`;
+          cardElement.style.left = `${(targetStack.length - 1) * cardShiftValue}px`;
+          if (targetStack.length > 1) {
+            targetStack.dom.style.width = `${targetStack.width + cardShiftValue * (targetStack.length - 1)}px`;
           }
           break;
         default:
