+++ src/js/display/playerTable/displayAside.js
@@ -232,13 +232,12 @@ const displayAside = {
     for (let i = 0; i < agesNumber; i += 1) {
       const deck = document.createElement('div');
       deck.classList.add('age-deck');
-      deck.id = `age${i + 1}`; // id age1-10 for each deck
+      deck.id = `age${i + 1}`;
       if (i < 9) {
         deck.style.backgroundImage = `url(./assets/img/cards-bg/age-0${i + 1}.jpg)`;
       } else if (i === 9) {
         deck.style.backgroundImage = 'url(./assets/img/cards-bg/age-10.jpg)';
       }
-
       deck.addEventListener('click', () => this.ageDecksBlock.classList.toggle('age-decks--hidden'));
 
       // Animation classes and attributes for cards here
