+++ src/js/z_temp/getCard.js
