+++ src/scss/abstracts/_mixins.scss
@@ -12,12 +12,12 @@
 
 @mixin header-active-card {
   border: solid 0.1rem #000;
+  border-radius: 0.4rem;
   font-size: 1.8rem;
   height: 2.2rem;
   text-align: center;
-  width: 3.8rem;
-  border-radius: 0.4rem;
   transition: all 0.5s;
+  width: 3.8rem;
 }
 
 @mixin block-background-overlay {
@@ -32,14 +32,14 @@
 @mixin title-settings($url, $size: cover) {
   background: url($url);
   background-size: $size;
-  width: 150px;
-  height: 62px;
-  position: absolute;
-  left: 50%;
-  transform: translateX(-50%);
-  top: -25px; // don't change on rem
-  z-index: 3;
   filter: drop-shadow(1px 2px 3px rgba(0, 0, 0, 0.8));
+  height: 62px;
+  left: 50%;
+  position: absolute;
+  top: -25px; // don't change on rem
+  transform: translateX(-50%);
+  width: 150px;
+  z-index: 3;
 }
 
 @mixin area-border($url) {
