+++ src/scss/base/_base.scss
@@ -15,10 +15,10 @@ html {
 }
 
 body {
+  font-family: 'Fira Sans Condensed';
   font-size: 2rem;
   height: auto;
   width: 100vw;
-  font-family: 'Fira Sans Condensed';
 
   ::-webkit-scrollbar {
     height: 0.5rem;
