+++ src/scss/components/aside/_currentDeck.scss
@@ -5,36 +5,36 @@
 }
 
 .current-deck__title {
-  height: 3rem;
-  width: 20rem;
   align-items: flex-end;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
+  background: url('../../assets/img/cards-bg/get-card.png') right no-repeat;
+  background-size: contain;
   border-radius: 1rem;
+  color: #fff;
+  display: flex;
+  filter: drop-shadow(2px 4px 6px rgb(0, 0, 0));
+  flex-direction: column;
   font-size: 2.1rem;
   font-weight: 700;
-  margin: 0;
-  text-align: right;
-  background: url('../../assets/img/cards-bg/get-card.png') right no-repeat;
-  z-index: 5;
-  position: absolute;
+  height: 3rem;
   height: 4rem;
-  background-size: contain;
-  top: 2.7rem;
+  justify-content: center;
   left: 10rem;
-  color: #fff;
+  margin: 0;
   padding-right: 1rem;
-  filter: drop-shadow(2px 4px 6px black);
+  position: absolute;
+  text-align: right;
   text-shadow: 1px 1px #000;
+  top: 2.7rem;
+  width: 20rem;
+  z-index: 5;
 }
 
 .current-deck__cards {
   @include flex(column, center, center);
   @include size(30rem, 20rem);
+  filter: drop-shadow(2px 4px 6px rgb(0, 0, 0));
   margin: 0;
   transition: 0.3s linear;
-  filter: drop-shadow(2px 4px 6px rgb(0, 0, 0));
 
   &:hover .hover-btn {
     opacity: 1;
