+++ src/scss/components/aside/_extraCards.scss
@@ -12,13 +12,13 @@
 .extra-cards__special-title {
   @include size(100%, 3rem);
   @include flex(column, center, center);
-  font-size: 2.5rem;
-  text-align: center;
-  position: absolute;
+  bottom: 3.2rem;
   color: #fff;
   font-family: 'Caesar';
-  bottom: 3.2rem;
-  text-shadow: 2px 2px 3px black;
+  font-size: 2.5rem;
+  position: absolute;
+  text-align: center;
+  text-shadow: 2px 2px 3px rgb(0, 0, 0);
   z-index: 2;
 }
 
@@ -29,7 +29,7 @@
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
-  filter: drop-shadow(2px 4px 6px black);
+  filter: drop-shadow(2px 4px 6px rgb(0, 0, 0));
 
   &:hover .hover-btn {
     opacity: 1;
@@ -45,13 +45,13 @@
 .extra-cards__leadership-title {
   @include size(100%, 3rem);
   @include flex(column, center, center);
-  font-size: 2.5rem;
-  text-align: center;
-  position: absolute;
+  bottom: 3.2rem;
   color: #fff;
   font-family: 'Caesar';
-  bottom: 3.2rem;
-  text-shadow: 2px 2px 3px black;
+  font-size: 2.5rem;
+  position: absolute;
+  text-align: center;
+  text-shadow: 2px 2px 3px rgb(0, 0, 0);
   z-index: 2;
 }
 
@@ -62,7 +62,7 @@
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
-  filter: drop-shadow(2px 4px 6px black);
+  filter: drop-shadow(2px 4px 6px rgb(0, 0, 0));
 
   &:hover .hover-btn {
     opacity: 1;
