+++ src/scss/layouts/_hand.scss
@@ -43,6 +43,7 @@
 }
 
 .hand__controls {
+  @include area-border('../../assets/img/border-hand-control.png');
   display: flex;
   flex-direction: column;
   height: 100%;
@@ -51,7 +52,6 @@
   right: -0.05rem;
   width: 6.5rem;
   z-index: 1;
-  @include area-border('../../assets/img/border-hand-control.png');
 }
 
 .hand__controls-svg {
