+++ src/scss/layouts/_header.scss
@@ -2,8 +2,8 @@
   @include flex(row, space-around, center);
   background: $tables-bg;
   height: 17rem;
-  width: calc(100vw - 37rem);
   position: relative;
+  width: calc(100vw - 37rem);
 
   .header__title {
     @include title-settings('../../assets/img/title-header.png');
@@ -17,32 +17,32 @@
   .header__table {
     @include area-border('../../assets/img/border-top.png');
     @include flex(row, center, flex-end);
-    width: 100%;
     height: 100%;
+    width: 100%;
 
     &__wrapper {
       @include flex(row, space-between, flex-end);
-      width: auto;
+      color: rgb(226, 226, 226);
       height: 100%;
       margin-left: 2.5rem;
       overflow-x: auto;
       position: relative;
+      width: auto;
       z-index: 2;
-      color: rgb(226, 226, 226);
     }
   }
 }
 
 .player-container {
   @include flex(column, space-between, center);
-  height: 14rem;
-  min-width: 30rem;
-  border-radius: 1rem;
   background: rgba(0, 0, 0, 0.3);
+  border-radius: 1rem;
   box-shadow: inset 0 0 6px 3px rgba(0, 0, 0, 0.5),
     0 0 3px 1px rgba(255, 255, 255, 0.1);
+  height: 14rem;
+  margin: 0 0.5rem 0.5rem;
+  min-width: 30rem;
   padding: 1rem;
-  margin: 0 0.5rem 0.5rem 0.5rem;
 
   &__hidden {
     display: none;
@@ -55,34 +55,34 @@
   width: 100%;
 
   &__name {
-    font-size: 2rem;
-    height: 100%;
-    text-align: center;
-    width: 40%;
+    @include flex;
     background: url('../../assets/img/player-container/player-title.png');
     background-size: cover;
-    font-family: 'Caesar';
-    @include flex;
-    padding-top: 0.5rem;
     color: rgb(255, 208, 0);
+    font-family: 'Caesar';
+    font-size: 2rem;
+    height: 100%;
+    padding-top: 0.5rem;
+    text-align: center;
+    width: 40%;
   }
 }
 
 .cards-container {
   @include flex(row, space-around, center);
-  height: 100%;
-  padding: 0.3rem;
-  width: 17%;
+  background: #3b1105;
   border: solid 0.1rem #ee8823;
   border-radius: 1rem;
-  background: #3b1105;
   box-shadow: inset 0 0 2px 2px rgba(0, 0, 0, 0.5),
     2px 0 3px 1px rgba(0, 0, 0, 0.5);
+  height: 100%;
+  padding: 0.3rem;
   transition: all 0.5s;
+  width: 17%;
 
   &:not(:last-child):hover {
-    transform: scale(1.1);
     cursor: pointer;
+    transform: scale(1.1);
   }
 
   &__icon {
@@ -100,8 +100,8 @@
   width: 100%;
 
   div:hover {
-    transform: scale(1.2);
     cursor: pointer;
+    transform: scale(1.2);
   }
 
   &__red {
@@ -138,14 +138,14 @@
 
 .recource-container {
   @include flex(column, space-around, center);
+  background: #3b1105;
+  border: solid 0.1rem #ee8823;
+  border-radius: 1rem;
+  box-shadow: inset 0 0 2px 2px rgba(0, 0, 0, 0.5),
+    2px 0 3px 1px rgba(0, 0, 0, 0.5);
   height: 100%;
   padding: 0.3rem;
   width: 15%;
-  border: solid 0.1rem #ee8823;
-  border-radius: 1rem;
-  background: #3b1105;
-  box-shadow: inset 0 0 2px 2px rgba(0, 0, 0, 0.5),
-    2px 0 3px 1px rgba(0, 0, 0, 0.5);
 
   &__icon {
     font-size: 1.6rem;
